requirements.txt
streamlit
requests
beautifulsoup4
pandas
mysql-connector-python
matplotlib